% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experimental_file.R
\name{add_column}
\alias{add_column}
\title{Complete data frame with columns}
\usage{
add_column(data, col_name, value = NULL)
}
\arguments{
\item{data}{a data frame of interest.}

\item{col_name}{a character. Name of column that should be created if it does
not exist}

\item{value}{optional. A value to fill with.}
}
\description{
This function adds column if does not exist and fill it with
provided value.
}
