% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise}
\alias{add_noise}
\title{Adds noise to mass domain}
\usage{
add_noise(curve, standard_deviations)
}
\arguments{
\item{curve}{`data.table`}

\item{standard_deviations}{named vector of standard deviations}
}
\value{
`data.table`
}
\description{
Adds noise to mass domain
}
\keyword{internal}
