% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_intensities}
\alias{add_noise_to_intensities}
\title{Adds noise to a single time point in intensity domain}
\usage{
add_noise_to_intensities(spectrum, standard_deviations)
}
\arguments{
\item{spectrum}{a single spectrum from the list obtained using
\code{\link[powerHaDeX]{make_experimental_design}}.}

\item{standard_deviations}{vector of standard deviations for the random noise}
}
\value{
data.table
}
\description{
Adds noise to a single time point in intensity domain
}
\details{
if the \code{intensity_deviations} were provided, then noise is
sampled from a normal distribution with mean \code{0} and standard deviation
equal to those deviations and added to \code{Intensity}.
}
\keyword{internal}
