% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_one_timepoint}
\alias{add_noise_to_one_timepoint}
\title{Adds noise to a single time point in mass domain}
\usage{
add_noise_to_one_timepoint(spectrum, standard_deviations)
}
\arguments{
\item{spectrum}{data.table with a single time point}

\item{standard_deviations}{vector of standard deviations for the random noise}
}
\value{
data.table
}
\description{
Adds noise to a single time point in mass domain
}
\keyword{internal}
