% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_spectra}
\alias{add_noise_to_spectra}
\title{Creates spectra with technical replicates and noise}
\usage{
add_noise_to_spectra(
  spectra,
  n_runs,
  n_replicates,
  undeuterated_mass,
  mass_deviations,
  intensity_deviations
)
}
\arguments{
\item{spectra}{list of spectra created by the `get_spectra_list` function}

\item{n_runs}{number of technical replicates to create}

\item{n_replicates}{number of replicates of a spectrum for power calculation}

\item{undeuterated_mass}{undeuterated mass of a peptide}

\item{mass_deviations}{mass deviation in parts per million. Either a single number
(then the error at each time point will be the same) or a vector of the same length
as number of unique time points in the experiment. The error will be sampled from
normal distribution with standard deviation equal to `mass_deviations`*`undeuterated_mass`/1e6}

\item{intensity_deviations}{optional, standard deviations of random noise that
will be added to intensities. Either a single number (then the error at each
time point will be the same) or a vector of the same length as number of unique
time points in the experiment. The error will be sampled from normal distribution
with these standard deviations.}
}
\value{
list of lists of data.tables
}
\description{
Creates spectra with technical replicates and noise
}
\keyword{internal}
