% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{auc_test}
\alias{auc_test}
\title{Test based on area under the deuteration curve}
\usage{
auc_test(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table of deuteration curves}

\item{significance_level}{significance level that will be used for testing}
}
\description{
Test based on area under the deuteration curve
}
