% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{calculate_hdx_power}
\alias{calculate_hdx_power}
\title{Calculate power of statistical tests for HDX experiments}
\usage{
calculate_hdx_power(
  deuteration_curves,
  tests,
  significance_level = 0.05,
  summarized = TRUE
)
}
\arguments{
\item{deuteration_curves}{list returned by the `get_noisy_deuteration_curves`}

\item{tests}{lists of tests to perform. Each test function should have}

\item{significance_level}{significance level that will be used for testing}

\item{summarized}{default \code{TRUE}}
}
\value{
list of data.tables with test result, optionally summarized with power
}
\description{
Calculate power of statistical tests for HDX experiments
}
