% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{calculate_peptide_mass}
\alias{calculate_peptide_mass}
\title{Calculate mass of undeuterated peptide}
\usage{
calculate_peptide_mass(sequence)
}
\arguments{
\item{sequence}{character vector of amino acid sequence of a peptide}
}
\description{
Calculate mass of undeuterated peptide
}
\keyword{internal}
