% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hadex_data.R
\name{convert_hadex_data}
\alias{convert_hadex_data}
\title{Convert HaDeX to powerHaDeX data type.}
\usage{
convert_hadex_data(dat)
}
\arguments{
\item{dat}{data.table. The data of hdx_data class from the HaDeX package.}
}
\value{
This function returns a data table..
}
\description{
This function converts the data from HaDeX in order to make it
compatible with the input of test functions.
}
\seealso{
The tests:

- \code{\link[powerHaDeX]{test_houde}},

- \code{\link[powerHaDeX]{test_hdx_analyzer}},

- \code{\link[powerHaDeX]{test_memhdx_model}},

-\code{\link[powerHaDeX]{test_semiparametric}}.

And \code{\link[powerHaDeX]{test_hadex_data}}.
}
\keyword{internal}
