% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplest_model.R
\name{deuteros}
\alias{deuteros}
\title{Simplest linear models for deuteration curves containing time, state of the protein and the interaction term,.}
\usage{
deuteros(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\description{
Simplest linear models for deuteration curves containing time, state of the protein and the interaction term,.
}
