% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{fix_columns_names_types}
\alias{fix_columns_names_types}
\title{Standardize column names and types}
\usage{
fix_columns_names_types(curves)
}
\arguments{
\item{curves}{list of lists of data.tables}
}
\description{
Standardize column names and types
}
