% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{get_approx_isotopic_distribution}
\alias{get_approx_isotopic_distribution}
\title{Approximates isotopic distribution via sampling}
\usage{
get_approx_isotopic_distribution(sequence, min_probability = 0.001)
}
\arguments{
\item{sequence}{amino acid sequence of a peptide}

\item{min_probability}{minimum isotopic probability that will be considered}
}
\value{
list
}
\description{
Approximates isotopic distribution via sampling
}
\keyword{internal}
