% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_deuteration_curve_single_spectrum}
\alias{get_deuteration_curve_single_spectrum}
\title{Get a deuteration curve from a single spectrum}
\usage{
get_deuteration_curve_single_spectrum(spectrum, relative)
}
\arguments{
\item{spectrum}{data.table with a spectrum}

\item{relative}{if TRUE (default), each curve will start at 0}
}
\value{
data.table
}
\description{
Get a deuteration curve from a single spectrum
}
\keyword{internal}
