% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_deuteration_curves_from_spectra}
\alias{get_deuteration_curves_from_spectra}
\title{Create a list of deuteration curves from a list of spectra}
\usage{
get_deuteration_curves_from_spectra(spectra)
}
\arguments{
\item{spectra}{list created by `get_spectra_list` and `add_noise_to_spectra`
functions.}
}
\value{
list of lists of data.tables
}
\description{
Create a list of deuteration curves from a list of spectra
}
\keyword{internal}
