% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_deuteration_single_timepoint}
\alias{get_deuteration_single_timepoint}
\title{Calculates deuteration for given timepoint}
\usage{
get_deuteration_single_timepoint(
  initial_matrix,
  time_sequence,
  hd_probs,
  dh_probs
)
}
\arguments{
\item{initial_matrix}{A matrix}

\item{time_sequence}{vector of exchange times}

\item{hd_probs}{probabilities of transition HD}

\item{dh_probs}{probabilities of transition DH}
}
\value{
a matrix denoting hydrogen-deuterium exchange for given timepoint.
}
\description{
Calculates deuteration for given timepoint
}
