% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_constants}
\alias{get_exchange_constants}
\title{Exchange constant for Hydrogen-Deuterium Exchange.}
\usage{
get_exchange_constants(pH, pkc_consts, k_consts)
}
\arguments{
\item{pH}{reaction pH}

\item{pkc_consts}{constants calculated via \code{\link[powerHaDeX]{get_pkc}}}

\item{k_consts}{constants calculated via
\code{\link[powerHaDeX]{get_poly_const}}}
}
\value{
a matrix named \code{constants} of tabular and calculated constants
(specifically for \code{Asp}, \code{Glu}, \code{His}, \code{C-Term} and
\code{NHMe})
}
\description{
Exchange constant for Hydrogen-Deuterium Exchange.
}
\keyword{internal}
