% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Hydrogen-deuterium or back-exchange exchange rates}
\usage{
get_exchange_rates(
  sequence,
  exchange = "HD",
  pH = 9,
  temperature = 15,
  mol_type = "poly",
  if_corr = 0
)
}
\arguments{
\item{sequence}{peptide amino acid sequence as a single string}

\item{exchange}{"HD" for hydrogen-deuterium exchange,
"DH" for deuterium-hydrogen (back-exchange).}

\item{pH}{pH of the reaction}

\item{temperature}{temperature of the reaction (Celsius)}

\item{mol_type}{... TODO: fill this in}

\item{if_corr}{pH correction factor}

\item{gas_constant}{gas constant}
}
\value{
numeric vector of exchange rates
}
\description{
Hydrogen-deuterium or back-exchange exchange rates
}
\keyword{internal}
