% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_HD_exchange.R
\name{get_HD_matrices_using_markov}
\alias{get_HD_matrices_using_markov}
\title{Get a matrix of simulated exchanged hydrogens for each experiment time point using markov chains}
\usage{
get_HD_matrices_using_markov(
  sequence,
  transition_probs,
  steps_between_time_points,
  n_molecules = 100
)
}
\arguments{
\item{sequence}{amino acid sequence of a peptide as a character vector}

\item{transition_probs}{list of probabilities of exchange returned by the get_exchange_probabilities function}

\item{steps_between_time_points}{A vector containing sum of steps between
times at which deuteration levels are measured.}

\item{n_molecules}{number of peptide molecules}
}
\value{
list of matrices
}
\description{
Get a matrix of simulated exchanged hydrogens for each experiment time point using markov chains
}
\keyword{internal}
