% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_HD_exchange.R
\name{get_intensity}
\alias{get_intensity}
\title{Get the isotopic probabilities for the deuterated peptide.}
\usage{
get_intensity(
  HD_matrices,
  maxD,
  maxND,
  isotopic_probs,
  peptide_mass,
  times,
  charge,
  pH
)
}
\arguments{
\item{HD_matrices}{list. Simulated matrices for every time point after hydrogen-deuterium-exchange}

\item{maxD}{length of the sequence - amount of prolines}

\item{maxND}{length of the isotopic distribution - 1}

\item{isotopic_probs}{the isotopic distribution for the undeuterated peptide.}

\item{peptide_mass}{mass of the peptide + mass of H2O}

\item{times}{times at which deuteration levels will be measured (seconds)}

\item{charge}{charge of the peptide ion.}

\item{pH}{pH of the reaction}
}
\description{
Compute the isotopic probabilities for the deuterated peptide as a convolution
of the isotopic distribution for the undeuterated peptide and the observed
isotopic distribution after hydrogen-deuterium exchange computed by \code{get_observed_iso_dist}.
}
