% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_noisy_deuteration_curves}
\alias{get_noisy_deuteration_curves}
\title{Create a list of lists of noisy deuteration curves from theoretical spectra}
\usage{
get_noisy_deuteration_curves(
  theoretical_spectra,
  compare_pairs = TRUE,
  reference = NA,
  n_runs = 4,
  n_replicates = 100,
  mass_deviations = 50,
  intensity_deviations = NULL,
  per_run_deviations = NULL,
  relative = TRUE
)
}
\arguments{
\item{theoretical_spectra}{data.table of spectra created by the
`simulate_theoretical_spectra` function}

\item{compare_pairs}{if FALSE, all groups (defined by the protection factor)
will be considered jointly. If TRUE (default), each protection factor will be
considered together with the protection factor given by the `reference` parameter}

\item{reference}{protection factor that will be used for comparison to other
protection factors in. The function accepts either \code{NA} (for comparing all
protection factors), a number (for comparing with reference value of protection factor)
or "all"(for pairwise comparisons of all the possible combinations). Default \code{NA}.}

\item{n_runs}{number of technical replicates to create}

\item{n_replicates}{number of replicates of a spectrum for power calculation}

\item{mass_deviations}{mass deviation in parts per million. Either a single number
(then the error at each time point will be the same) or a vector of the same length
as number of unique time points in the experiment. The error will be sampled from
normal distribution with standard deviation equal to `mass_deviations`*`undeuterated_mass`/1e6}

\item{intensity_deviations}{optional, standard deviations of random noise that
will be added to intensities. Either a single number (then the error at each
time point will be the same) or a vector of the same length as number of unique
time points in the experiment. The error will be sampled from normal distribution
with these standard deviations.}

\item{per_run_deviations}{optional, standard deviations of random noise that
will be added to deuteration curves. Either a single number (then the error at each
time point will be the same) or a vector of the same length as number of unique
time points in the experiment. The error will be sampled from normal distribution
with these standard deviations.}

\item{relative}{logical, if TRUE (default), each deuteration curve will start at 0}
}
\value{
list of lists of data.tables
}
\description{
Create a list of lists of noisy deuteration curves from theoretical spectra
}
