% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_HD_exchange.R
\name{get_observed_iso_dist}
\alias{get_observed_iso_dist}
\title{Get observed isotopic distribition after hydrogen-deuterium exchange}
\usage{
get_observed_iso_dist(HDmatrix, isotopic_distribution, maxD)
}
\arguments{
\item{HDmatrix}{simulated matrix after hydrogen-deuterium-exchange}

\item{isotopic_distribution}{vector of isotopic probabilities of a peptide}

\item{maxD}{length of the isotopic distribution - 1}
}
\description{
Get observed isotopic distribition after hydrogen-deuterium exchange
}
\keyword{internal}
