% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_poly_const}
\alias{get_poly_const}
\title{Constant related to...}
\usage{
get_poly_const(mol_type, exchange = "HD")
}
\arguments{
\item{mol_type}{TODO: fill this in}

\item{exchange}{type of exchange - "HD" for hydrogen to deuterium,
"DH" for deuterium to hydrogen (back-exchange).}
}
\value{
list
}
\description{
Constant related to...
}
\keyword{internal}
