% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_spectra_list}
\alias{get_spectra_list}
\title{Get a list of spectra}
\usage{
get_spectra_list(theoretical_spectra, compare_pairs, reference)
}
\arguments{
\item{theoretical_spectra}{data.table of spectra created by the
`simulate_theoretical_spectra` function}

\item{compare_pairs}{if FALSE, all groups (defined by the protection factor)
will be considered jointly. If TRUE (default), each protection factor will be
considered together with the protection factor given by the `reference` parameter}

\item{reference}{protection factor that will be used for comparison to other
protection factors in}
}
\value{
list of data.tables
}
\description{
Get a list of spectra
}
