% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{hdx_anova}
\alias{hdx_anova}
\title{ANOVA based on gls}
\usage{
hdx_anova(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table of deuteration curves at given time point}

\item{significance_level}{significance level for testing}
}
\description{
ANOVA based on gls
}
