% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{houde}
\alias{houde}
\title{Houde's test for deuteration curves}
\usage{
houde(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\value{
This function returns a data table compatible with the function
\code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\description{
This function performs Damian Houde's confidence intervals test
for differences in deuteration levels. Its input and output are compatible with
the function \code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\references{
Houde, Damian, Steven A Berkowitz, and John R Engen (2011).
“The utility of hydrogen/deuterium exchange mass spectrometry in biopharmaceutical
comparabilitystudies”. In:Journal of pharmaceutical sciences100.6, pp. 2071–2086.
}
\seealso{
\code{\link[powerHaDeX]{calculate_hdx_power}} for estimation of power
of tests for differences in deuteration levels.
}
