% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{lme_model}
\alias{lme_model}
\title{Linear mixed effects model proposed by Jurgen Claesen}
\usage{
lme_model(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table of deuteration curves}

\item{significance_level}{significance level for testing}
}
\description{
Linear mixed effects model proposed by Jurgen Claesen
}
