% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{make_experimental_design}
\alias{make_experimental_design}
\title{Create a list of spectra with replicates from a a list of spectra}
\usage{
make_experimental_design(spectra, n_runs)
}
\arguments{
\item{spectra}{list created by the `get_spectra_list` function}

\item{n_runs}{number of technical replicates in the experiment}
}
\value{
list of data.tables
}
\description{
Create a list of spectra with replicates from a a list of spectra
}
\keyword{internal}
