% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memhdx_model.R
\name{memhdx_model}
\alias{memhdx_model}
\title{Linear mixed effects model used in MEMHDX tools}
\usage{
memhdx_model(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table of deuteration curves}

\item{significance_level}{significance level for testing}
}
\description{
Linear mixed effects model used in MEMHDX tools
}
