% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_single_spectrum}
\alias{plot_single_spectrum}
\title{Draw mass spectrum}
\usage{
plot_single_spectrum(spectrum, time_point = NULL, ...)
}
\arguments{
\item{spectrum}{data table. Result of \code{\link[powerHDX]{simulate_theoretical_spectra}}}

\item{time_point}{a value of exposure for which plot will be provided. If \code{NULL}
exposure time will be taken from \code{spectrum} but only when there will be one value.
Default to \code{NULL}.}

\item{...}{additional arguments passing to the \code{\link[ggplot2]{theme}}.}
}
\value{
ggplot object
}
\description{
Draw mass spectrum
}
\details{
This function draws mass spectrum from data obtained via
\code{\link[powerHDX]{simulate_theoretical_spectra}}.
}
\examples{
\dontrun{
spectrum <-   simulate_theoretical_spectra(sequence = "AAHLPAEFTPAVHASLDKFLASVSTVLTSKYR",
              charge = 3,
              protection_factor = 100,
              times = 60,
              pH = 6,
              temperature = 15,
              n_molecules = 500,
              time_step_const = 1,
              use_markov = TRUE)
#one exposure in data so time_point can be NULL
plot_single_spectrum(spectrum)

spectrum <-   simulate_theoretical_spectra(sequence = "AAHLPAEFTPAVHASLDKFLASVSTVLTSKYR",
              charge = 3,
              protection_factor = 100,
              times = c(60, 600),
              pH = 6,
              temperature = 15,
              n_molecules = 500,
              time_step_const = 1,
              use_markov = TRUE)
#in case of two time points one must be chosen
plot_single_spectrum(spectrum, 60)
}

}
