% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerHaDeX-package.R
\docType{package}
\name{powerHaDeX-package}
\alias{powerHaDeX}
\alias{powerHaDeX-package}
\title{powerHaDeX: Power analysis for HDX-MS experiments}
\description{
The R-package powerHaDeX provides tools for simulating and analyzing data
coming from HDX-MS experiments, along with the possibility of comparing the
power of tests verifying differences in deuteration levels.
}
\author{
Michal Burdukiewicz, Krystyna Grzesiak, Mateusz Staniak
}
\keyword{internal}
