% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerHaDeX-package.R
\docType{package}
\name{powerHaDeX}
\alias{powerHaDeX}
\title{powerHaDeX}
\description{
The R-package powerHaDeX for simulating and analyzing data coming
from HDX-MS experiments along with the possibility of comparing the power of
the tests verifying differences in deuteration levels.
}
\author{
Michal Burdukiewicz, Krystyna Grzesiak, Mateusz Staniak
}
