% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experimental_file.R
\name{prepare_input_peptides}
\alias{prepare_input_peptides}
\title{Prepare input for \code{\link[powerHDX]{create_experimental_file}}}
\usage{
prepare_input_peptides(peptides)
}
\arguments{
\item{peptides}{a data frame of parameters for which \code{\link[powerHDX]{simulate_theoretical_spectra}}
will be executed.}
}
\value{
a data frame being a proper input for \code{\link[powerHDX]{create_experimental_file}}.
}
\description{
Supplementary function providing appropriate input.
}
