% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_models.R
\name{S2_by_4}
\alias{S2_by_4}
\title{Semiparametric for deuteration curves}
\usage{
S2_by_4(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\description{
Semiparametric for deuteration curves
}
