% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric.R
\name{semiparametric}
\alias{semiparametric}
\title{Semiparametric test for differences in deuteration}
\usage{
semiparametric(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\description{
This function performs the semiparametric test for differences
in deuteration levels. Its input and output are compatible with
the function \code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\details{
This function uses \code{\link[powerHaDeX]{truncated_lines}}. The knots
considered in the testing procedure are chosen using ridge regression.
}
