% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_theoretical_spectra.R
\name{simulate_theoretical_spectra}
\alias{simulate_theoretical_spectra}
\title{Simulate theoretical spectra of a deuterated peptide over time}
\usage{
simulate_theoretical_spectra(
  sequence,
  charge = NULL,
  protection_factor = 1,
  times = c(60, 600),
  pH = 7.5,
  temperature = 15,
  n_molecules = 100,
  time_step_const = 1,
  if_corr = 0,
  min_probability = 1e-04,
  use_markov = TRUE
)
}
\arguments{
\item{sequence}{amino acid sequence of a peptide as a single string}

\item{charge}{vector of charges of the peptide ion. If NULL, one value is sampled
from vector 2:6. Default NULL.}

\item{protection_factor}{protection factor. If a single number of provided,
same protection factor will be assumed for each amide. Default value: 1
(indicates that the exchange rate is equal to the intristic exchange rate)}

\item{times}{a vector of times at which deuteration levels will be measured (seconds)}

\item{pH}{pH of the reaction. Default to 7.5.}

\item{temperature}{temperature of the reaction (Celsius)}

\item{n_molecules}{number of peptide molecules. Default to 100.}

\item{time_step_const}{constant that will be dived by a maximum exchange rate
to obtain time step. Hydrogen-deuteration exchange may occur at each of these steps.}

\item{if_corr}{pH correction indicator. Default value 0. The value of pH is equal to pD.
If there is correction, the pD = pH + 0.4. (Conelly et al 1993)}

\item{min_probability}{smallest isotopic probability to consider}

\item{use_markov}{logical. If TRUE algorithm basing on Markov chain will be used.
If FALSE simulation provided by Zhong-Yuan Kan will be executed. Default to TRUE,
as it fastens the calculation}
}
\value{
data.table
}
\description{
Simulate theoretical spectra of a deuterated peptide over time
}
