% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{test_auc_test}
\alias{test_auc_test}
\title{Test based on area under the deuteration curve for differences in deuteration levels}
\usage{
test_auc_test(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\value{
This function returns a data table compatible with the function
\code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\description{
Test based on area under the deuteration curve for differences in deuteration levels
}
\seealso{
Mazur, Sharlyn J and Daniel P Weber (2017). “The area between exchange
curves as a measure of conformational differences in hydrogen-deuterium exchange
mass spectrometry studies”. In:Journal of the American Society for Mass Spectrometry
28.5, pp. 978–981.
}
