% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiparametric_models.R
\name{truncated_lines}
\alias{truncated_lines}
\title{Truncated lines}
\usage{
truncated_lines(x, kappa)
}
\arguments{
\item{x}{vector}

\item{kappa}{a number. Knot}
}
\description{
Truncated lines
}
