% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise}
\alias{add_noise}
\title{Adds noise to mass domain}
\usage{
add_noise(replicate_curve, standard_deviations)
}
\arguments{
\item{standard_deviations}{named vector of standard deviations}
}
\value{
`data.table` containing a single noisy curve (for one replication).
}
\description{
Adds noise to mass domain
}
\details{
The noise is sampled from normal distribution with mean \code{0}
and standard deviation equal to \code{per_run_deviations} and added to the
\code{Mass} values unless they are not zeroes (there is no noise at the time \code{0}).
}
\keyword{internal}
