% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_one_spectrum}
\alias{add_noise_to_one_spectrum}
\title{Add noise to a single spectrum}
\usage{
add_noise_to_one_spectrum(
  spectrum,
  undeuterated_mass,
  mass_deviations,
  intensity_deviations
)
}
\arguments{
\item{undeuterated_mass}{undeuterated mass of a peptide}

\item{mass_deviations}{mass deviation in parts per million. Either a single number
(then the error at each time point will be the same) or a vector of the same length
as number of unique time points in the experiment. The error will be sampled from
normal distribution with standard deviation equal to `mass_deviations`*`undeuterated_mass`/1e6}

\item{intensity_deviations}{optional, standard deviations of random noise that
will be added to intensities. Either a single number (then the error at each
time point will be the same) or a vector of the same length as number of unique
time points in the experiment. The error will be sampled from normal distribution
with these standard deviations.}
}
\value{
data.table
}
\description{
Add noise to a single spectrum
}
\keyword{internal}
