% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_one_timepoint}
\alias{add_noise_to_one_timepoint}
\title{Adds noise to a single time point in mass domain}
\usage{
add_noise_to_one_timepoint(spectrum, standard_deviations)
}
\arguments{
\item{spectrum}{a single spectrum from the list obtained using
\code{\link[powerHaDeX]{make_experimental_design}}.}

\item{standard_deviations}{vector of standard deviations for the random noise}
}
\value{
data.table
}
\description{
Adds noise to a single time point in mass domain
}
\details{
the noise is sampled from a normal distribution with mean \code{0}
and standard deviation equal to \code{standard_deviations} and added to
\code{Mz} values for the time points (based on supplied parameters for time
points).
}
\keyword{internal}
