% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{add_noise_to_single_curve}
\alias{add_noise_to_single_curve}
\title{Adds noise to a single deuteration curve}
\usage{
add_noise_to_single_curve(replicate_curve, per_run_deviations, relative)
}
\arguments{
\item{replicate_curve}{`data.table` with technical replicates of a deuteration curve}

\item{per_run_deviations}{vector of standard deviations of random error}
}
\value{
`data.table`
}
\description{
Adds noise to a single deuteration curve
}
\keyword{internal}
