% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{calculate_hdx_power}
\alias{calculate_hdx_power}
\title{Calculate power for a given set of tests and protection factors}
\usage{
calculate_hdx_power(
  sequence,
  protection_factors,
  tests,
  time_points,
  n_runs,
  n_replicates,
  mass_deviations,
  intensity_deviations = NULL,
  deuteration_deviations = NULL,
  ...
)
}
\arguments{
\item{sequence}{amino acid sequence of a peptide (as a single string).}

\item{protection_factors}{vector of protection factors that will be used in simulations.}

\item{tests}{a list of tests that will be evaluated.}

\item{time_points}{time points for simulation.}

\item{n_runs}{number of replicate runs of the mass spectrometer that will be simulated.}

\item{n_replicates}{number of replicates that will be used for power calculation.}

\item{mass_deviations}{a scalar or a vector of magnitudes of error in m/z dimension in ppm.}

\item{intensity_deviations}{standard deviations for isotopic probabilities, which
are used in place of intensities in simulations. If `NULL`, only error on the m/z dimension
will be used. If scalar, the same distribution will be used for each time point.
If a vector of parameters is provided, the random error will have a different variance
at each time point.}

\item{deuteration_deviations}{standard deviation for normal distribution that will be
used to model errors in deuteration curves. If `NULL`, only error on the m/z dimension
will be used. If scalar, the same distribution will be used for each time point.
If a vector of parameters is provided, the random error will have a different variance
at each time point.}

\item{...}{additional parameters passed to `simulate_theoretical_spectra` function.}
}
\value{
list
}
\description{
Calculate power for a given set of tests and protection factors
}
