% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{calculate_peptide_mass}
\alias{calculate_peptide_mass}
\title{Peptide mass}
\usage{
calculate_peptide_mass(sequence)
}
\arguments{
\item{sequence}{character vector of amino acid sequence of a peptide}
}
\value{
a single number denoting the mass of the undeuterated peptide.
}
\description{
Calculate mass of undeuterated peptide
}
\details{
Calculates peptide mass as a sum of amino acids' from
\code{sequence} masses and H2O mass (1.007825 * 2 + 15.994915 = 18.01056).
}
