% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_constants}
\alias{get_exchange_constants}
\title{Exchange constant for Hydrogen-Deuterium Exchange.}
\usage{
get_exchange_constants(pH, pkc_consts, k_consts)
}
\arguments{
\item{pH}{reaction pH}

\item{pkc_consts}{constants...}

\item{k_consts}{constants...}
}
\value{
matrix
}
\description{
Exchange constant for Hydrogen-Deuterium Exchange.
}
\keyword{internal}
