% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Hydrogen-deuterium or back-exchange exchange rates}
\usage{
get_exchange_rates(
  sequence,
  exchange = "HD",
  pH = 9,
  temperature = 15,
  mol_type = "poly",
  if_corr = 0
)
}
\arguments{
\item{sequence}{peptide amino acid sequence as a character vector of amino acids}

\item{exchange}{type of exchange - "HD" for hydrogen to deuterium,
"DH" for deuterium to hydrogen (back-exchange). Default "HD.}

\item{pH}{reaction pH}

\item{temperature}{temperature of the reaction (Celsius). Default to 15.}

\item{mol_type}{character, "poly" or "oligo".}

\item{if_corr}{pH correction indicator. Default value 0. The value of pH is equal to pD.
If there is correction, the pD = pH + 0.4. (Conelly et al 1993)}

\item{gas_constant}{gas constant}
}
\value{
numeric vector of exchange rates according to the provided exchange direction.
}
\description{
Hydrogen-deuterium or back-exchange exchange rates
}
\keyword{internal}
