% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_pkc}
\alias{get_pkc}
\title{Constant related to...}
\usage{
get_pkc(temp_kelvin, gas_constant, exchange = "HD")
}
\arguments{
\item{temp_kelvin}{temperature reaction in Kelvins}

\item{gas_constant}{gas constant (1/(dT * R))}

\item{exchange}{type of exchange - "HD" for hydrogen to deuterium,
"DH" for deuterium to hydrogen (back-exchange). Default "HD.}
}
\description{
This function calculates supplementary constants for aspartic
acid (Asp), glutamic acid (Glu) and histidine (His).
}
\keyword{internal}
