% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_pkc}
\alias{get_pkc}
\title{Constant related to...}
\usage{
get_pkc(temp_kelvin, gas_constant, exchange = "HD")
}
\arguments{
\item{temp_kelvin}{temperature reaction in Kelvins}

\item{gas_constant}{gas constant (1/(dT * R))}
}
\description{
Constant related to...
}
