% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_relative_mass}
\alias{get_relative_mass}
\title{Get relative mass}
\usage{
get_relative_mass(mass, time)
}
\arguments{
\item{mass}{vector of masses}

\item{time}{vector of time points}
}
\description{
Mass minus mass at zero is calculated as a form of deuterium
uptake.
}
\keyword{internal}
