% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{get_undeuterated_mass}
\alias{get_undeuterated_mass}
\title{Get mass of an undeuterated peptide}
\usage{
get_undeuterated_mass(theoretical_spectra)
}
\arguments{
\item{theoretical_spectra}{`data.table` returned by the `simulate_theoretical_spectra` function.}
}
\value{
`data.table`
}
\description{
Get mass of an undeuterated peptide
}
\keyword{internal}
