% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houde.R
\name{houde}
\alias{houde}
\title{Houde's test for deuteration curves}
\usage{
houde(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\description{
Houde's test for deuteration curves
}
