% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy_spectra.R
\name{make_experimental_design}
\alias{make_experimental_design}
\title{Create a list of spectra with replicates from a list of spectra}
\usage{
make_experimental_design(spectra, n_replicates = 4)
}
\arguments{
\item{spectra}{list created by the \code{\link[powerHaDeX]{get_spectra_list}}
function}

\item{n_replicates}{number of technical replicates to create}
}
\value{
list of data.tables
}
\description{
This function is used to prepare technical replicates from the
provided spectra. It is a supplementary function used in
\code{\link[powerHaDeX]{add_noise_to_spectra}}.
}
\keyword{internal}
