% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{memhdx_model}
\alias{memhdx_model}
\title{MEMHDX model}
\usage{
memhdx_model(data, significance_level = 0.05)
}
\arguments{
\item{data}{data.table with deuteration curves}

\item{significance_level}{significance level for tests}
}
\value{
This function returns a data table compatible with the function
\code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\description{
This function performs the test based on a linear mixed effects
model used in MEMHDX tools. Its input and output are compatible with the function
\code{\link[powerHaDeX]{calculate_hdx_power}}.
}
\seealso{
Hourdel, Véronique et al. (July 2016). “MEMHDX: an interactive tool
to expedite thestatistical validation and visualization of large HDX-MS data sets”.
In:Bioinformatics32.22, pp. 3413–3419.issn: 1367-4803.
}
