% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_theoretical_spectra.R
\name{simulate_theoretical_spectra}
\alias{simulate_theoretical_spectra}
\title{Simulate theoretical spectra of a deuterated peptide over time}
\usage{
simulate_theoretical_spectra(
  sequence,
  charge = NULL,
  protection_factor = 1,
  times = c(60, 600),
  pH = 7.5,
  temperature = 15,
  n_molecules = 100,
  time_step_const = 1,
  if_corr = 0,
  min_probability = 1e-04,
  use_markov = TRUE
)
}
\arguments{
\item{sequence}{amino acid sequence of a petpide as a single string}

\item{charge}{charge of the peptide ion.}

\item{protection_factor}{protection factor. If a single number of provided,
same protection factor will be assumed for each amide.}

\item{times}{times at which deuteration levels will be measured (seconds)}

\item{pH}{pH of the reaction}

\item{temperature}{temperature of the reaction (Celsius)}

\item{n_molecules}{number of peptide molecules}

\item{time_step_const}{constant that will be dived by a maximum exchange rate
to obtain time step. Hydrogen-deuteration exchange may occur at each of these steps.}

\item{if_corr}{correction factor for pH - 0 or 1}

\item{min_probability}{smallest isotopic probability to consider}

\item{use_markov}{logical. If TRUE algorithm basing on Markov chain will be used.
If FALSE simulation from kanzy??? will be executed.}
}
\value{
data.table
}
\description{
Simulate theoretical spectra of a deuterated peptide over time
}
