% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{truncated_lines}
\alias{truncated_lines}
\title{Get truncated lines}
\usage{
truncated_lines(x, knots)
}
\arguments{
\item{x}{a vector. Variable which undergoes the truncating}

\item{knots}{a vector of knots in which the spline model will break}
}
\description{
This function prepares the function space for semiparametric
model using truncated lines.
}
\keyword{internal}
